add not found service hook to perform additional cleanup
We use this in Cilium to synchronize the EndpointSlice in a clustermesh.
Since this behavior is controller by an annotation we could be in a
case where a user remove this annotation while the service still exist.
In that case we must perform additional cleanup of the EndpointSlice
since they won't be deleted by the normal owner reference mechanism.

Signed-off-by: Arthur Outhenin-Chalandre <arthur@cri.epita.fr>